rename to ensureBoardFormat
